+++ Editor/FloatVariableEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Variables.Editor
+{
+    [CustomEditor(typeof(FloatVariable), true)]
+    public class FloatVariableEditor : RuntimeVariableEditor<float> { }
+}