+++ Editor/IntVariableEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Variables.Editor
+{
+    [CustomEditor(typeof(IntVariable), true)]
+    public class IntVariableEditor : RuntimeVariableEditor<int> { }
+}