+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - Unreleased
+
+### Added
+
+- `RuntimeVariableEditor<T>` abstract base class
+- `BoolVariableEditor` custom editor
+- `DoubleVariableEditor` custom editor
+- `FloatVariableEditor` custom editor
+- `IntVariableEditor` custom editor
+- `StringVariableEditor` custom editor
+- `TransformVariableEditor` custom editor
+
 ## [0.3.0] - 2022-11-30
 
 ### Added